Fixed db issues, security issues, added random name gen
