+++ dpc-go/dpc-attribution/src/repository/implementer_org_repository_test.go
@@ -34,11 +34,11 @@ func (suite *ImplementerOrgRepositoryTestSuite) TestFindRelation() {
     expectedQuery := "SELECT id, implementer_id, organization_id, created_at, updated_at, deleted_at, status FROM implementer_org_relation WHERE implementer_id = \\$1 AND organization_id = \\$2"
 
 	rows := sqlmock.NewRows([]string{"id", "implementer_id", "organization_id", "created_at", "updated_at", "deleted_at", "status"}).
-        AddRow(suite.fakeRel.ID, suite.fakeRel.Implementer_ID, suite.fakeRel.Organization_ID,suite.fakeRel.CreatedAt, suite.fakeRel.UpdatedAt, nil,suite.fakeRel.Status)
+        AddRow(suite.fakeRel.ID, suite.fakeRel.ImplementerID, suite.fakeRel.OrganizationID,suite.fakeRel.CreatedAt, suite.fakeRel.UpdatedAt, nil,suite.fakeRel.Status)
 
-	mock.ExpectQuery(expectedQuery).WithArgs(suite.fakeRel.Implementer_ID,suite.fakeRel.Organization_ID).WillReturnRows(rows)
+	mock.ExpectQuery(expectedQuery).WithArgs(suite.fakeRel.ImplementerID,suite.fakeRel.OrganizationID).WillReturnRows(rows)
 
-	rel, err := repo.FindRelation(ctx, suite.fakeRel.Implementer_ID,suite.fakeRel.Organization_ID)
+	rel, err := repo.FindRelation(ctx, suite.fakeRel.ImplementerID,suite.fakeRel.OrganizationID)
 	assert.NoError(suite.T(), err)
 	assert.Equal(suite.T(), suite.fakeRel.ID, rel.ID)
 }
@@ -54,9 +54,9 @@ func (suite *ImplementerOrgRepositoryTestSuite) TestFindNonExistentRelation() {
     rows := sqlmock.NewRows([]string{"id", "implementer_id", "organization_id", "created_at", "updated_at", "deleted_at", "status"})
     //Did not add rows here, simulate missing record
 
-    mock.ExpectQuery(expectedQuery).WithArgs(suite.fakeRel.Implementer_ID,suite.fakeRel.Organization_ID).WillReturnRows(rows)
+    mock.ExpectQuery(expectedQuery).WithArgs(suite.fakeRel.ImplementerID,suite.fakeRel.OrganizationID).WillReturnRows(rows)
 
-    Implementer, err := repo.FindRelation(ctx, suite.fakeRel.Implementer_ID,suite.fakeRel.Organization_ID)
+    Implementer, err := repo.FindRelation(ctx, suite.fakeRel.ImplementerID,suite.fakeRel.OrganizationID)
 	assert.Error(suite.T(), err)
 	assert.Empty(suite.T(), Implementer)
 }
@@ -70,11 +70,11 @@ func (suite *ImplementerOrgRepositoryTestSuite) TestInsert() {
 	expectedInsertQuery := "INSERT INTO implementer_org_relation \\(implementer_id, organization_id, status\\) VALUES \\(\\$1, \\$2, \\$3\\) returning id, implementer_id, organization_id, created_at, updated_at, deleted_at, status"
 
     rows := sqlmock.NewRows([]string{"id", "implementer_id", "organization_id", "created_at", "updated_at", "deleted_at", "status"}).
-		AddRow(suite.fakeRel.ID, suite.fakeRel.Implementer_ID, suite.fakeRel.Organization_ID,suite.fakeRel.CreatedAt, suite.fakeRel.UpdatedAt, nil,suite.fakeRel.Status)
+		AddRow(suite.fakeRel.ID, suite.fakeRel.ImplementerID, suite.fakeRel.OrganizationID,suite.fakeRel.CreatedAt, suite.fakeRel.UpdatedAt, nil,suite.fakeRel.Status)
 
-	mock.ExpectQuery(expectedInsertQuery).WithArgs(suite.fakeRel.Implementer_ID,suite.fakeRel.Organization_ID, suite.fakeRel.Status).WillReturnRows(rows)
+	mock.ExpectQuery(expectedInsertQuery).WithArgs(suite.fakeRel.ImplementerID,suite.fakeRel.OrganizationID, suite.fakeRel.Status).WillReturnRows(rows)
 
-	rel, err := repo.Insert(ctx, suite.fakeRel.Implementer_ID, suite.fakeRel.Organization_ID, suite.fakeRel.Status)
+	rel, err := repo.Insert(ctx, suite.fakeRel.ImplementerID, suite.fakeRel.OrganizationID, suite.fakeRel.Status)
 	assert.NoError(suite.T(), err)
 	assert.Equal(suite.T(), suite.fakeRel.ID, rel.ID)
 }
